requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
google-generativeai